db-py/requirements.txt
mysql-connector-python >= 8.0
dask >= 2.8
sqlalchemy >= 1.3
